// src/platform/plugins/shared/data/server/search/types.d.ts
export interface SearchStrategyDependencies {
    savedObjectsClient: SavedObjectsClientContract;
    esClient: IScopedClusterClient;
    uiSettingsClient: Pick<IUiSettingsClient, 'get'>;
    searchSessionsClient: IScopedSearchSessionsClient;
    request: KibanaRequest;
    rollupsEnabled?: boolean;
}
// ...
export interface ISearchStrategy<SearchStrategyRequest extends IKibanaSearchRequest = IEsSearchRequest, SearchStrategyResponse extends IKibanaSearchResponse = IEsSearchResponse> {
    search: (request: SearchStrategyRequest, options: ISearchOptions, deps: SearchStrategyDependencies) => Observable<SearchStrategyResponse>;
    cancel?: (id: string, options: ISearchOptions, deps: SearchStrategyDependencies) => Promise<void>;
    extend?: (id: string, keepAlive: string, options: ISearchOptions, deps: SearchStrategyDependencies) => Promise<void>;
}